proyecto-main/requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.11.0
django-admin-interface==0.28.6
django-extra-settings==0.12.0
django-import-export==3.3.8